Move room back into Sockets folder
